aeternum/backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
supabase
anthropic
opencv-python
ultralytics
torch
torchvision
python-dotenv
Pillow
numpy
